% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{get_data_urls}
\alias{get_data_urls}
\title{get_data_urls}
\usage{
get_data_urls(sf, base_url = "ANY")
}
\arguments{
\item{sf}{a summary file object.}

\item{base_url}{the base URL for the data files. Default value is last known
location at the time this version of the package was prepared.}
}
\value{
A \code{tibble}.
}
\description{
Get URLs of data for the given summary file.
}
