% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{get_filename_patterns}
\alias{get_filename_patterns}
\title{get_filename_patterns}
\usage{
get_filename_patterns(sf)
}
\arguments{
\item{sf}{a summary file object.}
}
\value{
A \code{tibble}.
}
\description{
Get patterns which can be used to identify geo files and data files relevant
to the given summary file.
}
