% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{interpret_data_filenames}
\alias{interpret_data_filenames}
\title{interpret_data_filenames}
\usage{
interpret_data_filenames(sf, paths)
}
\arguments{
\item{sf}{a summary file object.}

\item{paths}{a vector of paths to data files corresponding to \code{sf}.}
}
\value{
A \code{tibble}.
}
\description{
Parse encoded filenames into a \code{tibble} which can be more easily
viewed and processed.
}
