% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\name{read_geo,SF2_2010,character-method}
\alias{read_geo,SF2_2010,character-method}
\title{read_geo}
\usage{
\S4method{read_geo}{SF2_2010,character}(sf, path)
}
\description{
Read a geo file from the 2010 SF2 summary file into a \code{tibble}. See
documentation for generic.
}
