% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{read_geo}
\alias{read_geo}
\title{read_geo}
\usage{
read_geo(sf, path)
}
\arguments{
\item{sf}{a summary file object.}

\item{path}{path to a geo file corresponding to \code{sf}.}
}
\value{
A \code{tibble} with the contents of the geo file.
}
\description{
Read a geo file into a \code{tibble}.
}
