% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SF2_2010-class}
\alias{SF2_2010-class}
\title{An S4 class to represent the 2010 SF2 summary file}
\description{
An S4 class to represent the 2010 SF2 summary file
}
\section{Slots}{

\describe{
\item{\code{type}}{Type of summary file ("SF2_2010")}
}}

