% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\docType{data}
\name{sf2_2010_geo_cols}
\alias{sf2_2010_geo_cols}
\title{SF2 2010 Geo File Columns}
\format{
A data frame with 101 rows and 6 columns:
\describe{
  \item{ID}{Unique ID for the field.}
  \item{FIELD}{Name of the field.}
  \item{START_POS}{Starting position of the field.}
  \item{FIELD_SIZE}{Size of the field.}
  \item{DATA_TYPE}{Type of the data contained in the field.}
  \item{DESCRIPTION}{Description of the geo field.}
}
}
\source{
\url{https://www.census.gov/data/datasets/2010/dec/summary-file-2.html}
}
\usage{
sf2_2010_geo_cols
}
\description{
A dataset which describes the fixed width file format.
}
\keyword{datasets}
