% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\docType{data}
\name{sf2_2010_iterations}
\alias{sf2_2010_iterations}
\title{SF2 2010 Characteristic Iterations}
\format{
A data frame with 331 rows and 2 columns:
\describe{
  \item{CODE}{Code used in the summary file data.}
  \item{DESCRIPTION}{Description of the characteristic iteration.}
}
}
\source{
\url{https://www.census.gov/data/datasets/2010/dec/summary-file-2.html}
}
\usage{
sf2_2010_iterations
}
\description{
A dataset with definitions of characteristic iterations (CHARITERs).
}
\keyword{datasets}
