% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\docType{data}
\name{sf2_2010_states}
\alias{sf2_2010_states}
\title{SF2 2010 States}
\format{
A data frame with 53 rows and 3 columns:
\describe{
  \item{NAME}{Name of the state.}
  \item{FIPS}{FIPS code associated with the state (or \code{NA} for the
  national level).}
  \item{ABBREV}{Name of the variable (or \code{US} for the national level).}
}
}
\source{
\url{https://www.census.gov/data/datasets/2010/dec/summary-file-2.html}
}
\usage{
sf2_2010_states
}
\description{
A dataset containing state (and US) information for which summary files are
defined. Names and abbreviations correspond to those in the files.
}
\keyword{datasets}
