% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\docType{data}
\name{sf2_2010_tables}
\alias{sf2_2010_tables}
\title{SF2 2010 Tables}
\format{
A data frame with 71 rows and 3 columns:
\describe{
  \item{TABLE}{Name of the table.}
  \item{DESCRIPTION}{Description of the table.}
}
}
\source{
\url{https://www.census.gov/data/datasets/2010/dec/summary-file-2.html}
}
\usage{
sf2_2010_tables
}
\description{
A dataset containing logical tables defined in the summary files.
}
\keyword{datasets}
