% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sfreader}
\alias{sfreader}
\alias{sfreader-package}
\title{sfreader: Census Summary File Reader}
\description{
An R Package to read Census summary files. At least some particular parts we need.
}
