% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2_2010.R
\docType{data}
\name{sf2_2010_sumlev}
\alias{sf2_2010_sumlev}
\title{SF2 2010 Summary Levels}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 145 rows and 5 columns.
}
\source{
\url{https://www.census.gov/data/datasets/2010/dec/summary-file-2.html}
}
\usage{
sf2_2010_sumlev
}
\description{
A dataset containing summary levels for state-specific files.
}
\details{
\describe{
  \item{CODE}{Code of the summary level.}
  \item{PARENT_CODE}{Name of the table.}
  \item{DESCRIPTION}{Description of the summary level.}
  \item{TYPE}{\code{NATIONAL} or \code{STATE}.}
  \item{GQ}{Boolean; \code{TRUE} if level applies to GQs.}
}


Summary levels specify geography levels of the data. They follow a
hierarchical structure, which can be accessed using the \code{PARENT_CODE}
variable. A value of \code{NA} indicates the summary level is at the top
level of the hierarchy. Use the \code{TYPE} and \code{GQ} fields to select
the correct hierarchy corresponding to state-specific or national files,
and whether group quarters data are in use.
}
\keyword{datasets}
